Moved more widgets out of exit bottom sheet
format
